Avoid null checks for view.getTokens() and instead use view.isUsingTokenView()
In the future, this could enable a complete separation of whether tokens are available vs whether we actually want a
global view. For now, it makes the calling code more self-descriptive as the logic is now clearly about whether we want
a global view and not so much whether there happens to be tokens available.