Simplify ZoneRenderer by not caching the fog buffer contents
Preserving the fog buffer no longer carries much benefit as the expensive part (calculated the exposed area) is now
cached on its own, and the rendering itself is typically quite cheap. And since the buffer contents have to be
invalidated during the key points of user interaction (e.g., moving the map), caching the contents does not offer a
perceivable benefit to the user.