Clean up after GetInfoFunctionTest
One of the tests was injecting a dummy implementation, but didn't clean up after itself. This made the order of the
tests important. But now we always restore a real implementation after the tests are done, so that tests can now be run
in any order and even multiple times.