Render each aura range separately, as is done for normal lights
Auras now calculate a separate area for each range, and associate those ranges with the corresponding colour in a
`DrawableLight`. Colourless ("bright") auras do not get a `DrawableLight` so that there is no obscuring of what lies
beneath. This makes multi-coloured auras meaningful and brings them closer to the way that normal lights are rendered.

`ZoneView.getLights(Type)` has now been specialized to just handle auras, since that is all it was being used for
anyways. Normal lights are handled a different way altogether.