Refactor lighting logic to split global vs personal illumination
At the core of this change is the addition of `Illuminator` and its result type `Illumination`. An `Illuminator` holds
all areas lit by normal lights, and associates them with their lumens value (this is actually done per sight multiplier
since that can change the light areas). It can then produce an `Illumination` that holds the various lumens levels as
unioned areas, from which we can calculate some derivate results such as which areas are lit and which areas are covered in
darkness. The `Illuminator` is also able to add and remove individual areas and will only recalculate the effected
lumens level.

The caching in `ZoneView` has also been reworked so that more information can be kept as things change. The main
lighting results are cached per `IlluminationKey` (which is just a sight multiplier for now) and which can be reused for
every token involved in a view. Other results are cached per `PlayerView`. This includes some results that were
previously specific to the current `PlayerView` being rendered, but now are stored in a `Map` so that they can still be
reused across view changes. As a result, view changes no longer require flushing the `ZoneView`, and correspondingly the
`zoneView.flush()` call has bee removed from `ZoneRenderer.invalidateCurrentViewCache()`.