Add a lumens overlay and make lights environmental
The new lumens overlay is a visualization of the light strength at each point of a map. The strength of the light is
determined by its lumens and - as always - darkness obscures light of equal or lesser lumens. The lumens overlay treats
lumens less than zero as darkness and renders dark areas as black. On the other end, lumens of 100 or greater are
treated as bright areas, and are rendered as a nearly transparent white. The final special case is lumens equal to zero,
which is a representation of daylight and is not included in the lumens overlay. Any other lumens values are drawn as
shades of grey, with low levels being nearly black, and with higher levels being closer to white.

The environmental lights are an improvement to the aesthetics of lights. Lights now blend with one another
"additively" (technically according to the Screen blend mode) so that they get brighter when they overlap. After
blending the lights, the results are blended with the map using a blend operation that brightens the map while
preserving details (similar to overlay blend mode and soft lighting). This make lights look more like they illuminate
the map rather than sit on top of it.

Both of these features can be toggled in the *View* menu depending on user preferences and game requirements. The lumens
overlay can also have its opacity configured in the Preferences dialog, and the older light and darkness opacities have
been removed. Lights no longer have an opacity congiguration since they are conceptually part of the map and not a layer
rendered on top.

In order to make the lumens overlay as useful as possible, the lumens values are now atached to individual light ranges
rather than to entire light sources. Existing campaigns will have their light sources' lumens values copied to each
range. The **Campaign Properties** dialog's _Light_ tab represents per-range lumens as by appending the lumens after the
optional range color. So these are now valid ranges:
```
r#ff0000+100
r#aa0000+50
r+100
```
These define, respectively, a saturated red range at 100 lumens, a desaturated red range at 50 lumens, and a colourless
range at 100 lumens.