Move selection logic into a dedicated model
This fixes two bugs where client selections were being cleared due to other clients moving tokens, and where the
selection history was having redundant entries while missing others.

The bulk of the change is to pull the selection logic out of `ZoneRenderer` and into a new `SelectionModel`. The
`SelectionModel` is solely responsible for maintain selections and the selection undo history. A new `SelectionChanged`
event allows components to react to selection changes without having to be explicitly notified.