Fall back to the map name if no player alias is set
The behaviour of these function is back to their 1.12.2 behaviour, as they will no longer return `null` values for maps
that have no player alias set:
- `getMapDisplayName()`
- `getAllMapDisplayNames()`
- `getVisibleMapDisplayNames()`

Also `getMapName()` will once again be able to find maps by matching the name when no player alias is set.

The new method `Zone.getDisplayName()` can be used to get the name of the zone to show to players, whether that is the
player alias or the regular name.