Catch JSON Exception when connecting to server
When trying to connect to a server, the application will attempt to parse the
return string as a JSON object. However, if the server doesn't exist, a `null`
string is returned. When trying to convert this to JSON, there's an
`IllegalStateException` and then the error pops up to the user.

After adding this body check, the application is able to resolve the error
gracefully. It will pop up a short dialog saying "Server "x" is not found" or
some such.