Change the lumens overlay for more consistent rendering
We now render only disjoint lumens levels rather than relying on `AlphaComposite.Src` to erase prior rendering. Using
the disjoint areas also means the boundaries can be draw overtop the various lumens levels rather than being
interleaved. The boundaries are also now rendered as black regardless of the overlay opacity, and the stroke has been
changed to use rounded corners as they are a little more consistent.

The `Illumination` class has been refactored a bit to keep it as focused as possible. It no longer exposes the base list
of `LumensLevel` that it was created from, but instead returns only those results that are meaningful to
callers (obscured lumens levels and disjoint obscured lumens levels). This also involved moving the logic for
incorporating extra lights (e.g., personal lights) out of `ZoneView` and into `Illumination`. The documentation has also
been increased to clarify the meaning of each result, as they are similar but distinct.