Add a per-map option to render lights with alpha compositing
Lights will still be blending with each other using the screen blend mode, but how the results are blended with the map
is now configurable. The new map option "Lighting Style" can be set to "Overlay" to render the results via alpha
compositing over that map, or "Environment" to render the results as an illumination of the map.

To support the "Overlay" style, the light overlay opacity was reintroduced to the Preferences dialog. The tooltip
mentions that this preferences only affect the "Overlay" style.