Fix light rendering so that GMs can see the color of darkness
Now that darkness and light exist in a single overlay, we ought not clip the rendering to only visible areas when
viewing as a GM. This was causing darkness to not actually be rendered, but only soft FoW was shown for GMs. Now that
clipping is only enabled for player views, GMs are once again shown the full nature of the darkness.