Update light and sight help text in the Campaign Properties dialog
The new syntax for per-range lumens no longer matched the existing help text. So the help text for has no been updated to
match the new syntax for defining sights and lights. The special cone option `offset=` was also missing from the help
text and has now been added.

The options for both sight and light are now in a description list to make them more readable and more consistent with
one another.