Handle nullabilty when converting LookupTable to DTO
Each reference field in `LookupTable` and `LookupEntry` is marked as `@Nullable` or `@Nonnull` for clarity. `@Nonnull`
is enforced for deserialized fields in `readResolve()`, with corresponding `null` checks in accessors being
removed. Several `Boolean` accessors have been changed to `boolean` since nullability for these fields in an internal
detail.

In `toDto()`/`fromDto()`, all `@Nullable` fields are checked first whether a value exists prior to converting.

The `visible` and `allowLookup` flags are now included in the copy constructor so that copied tables have the same flags
set.