Tweak the toolbar separators
All separators in the toolbar are now created via a helper method that builds a separator with symmetric padding. This
is done in favour of sandwiching a separator between struts and allows the separator itself to decide how to pad
itself. If orientation support is added again in the future, the separator will provide a way to pad itself
appropriately with needing separator struts for vertical and horizontal orientations.

As a result, the separator between the tools and the option panel now has symmetric padding instead of being right up
against the VBL tool button. A separator was also added in the *BL option panel, between the shape tools and the mode
selector. This distinguishes the two now that we have a flatter and less colourful set of icons.

The toolbar is no longer floatable. This happened to be the case already since 1.12 due to FlatLaf defaulting to
non-floatable toolbars. By making this explict, we don't have to keep around the logic needed for floating and
orientation changes, nor will this be a theme-dependent bit of functionality. Floating can always be added again in the
future if it's deemed desirable, along with better orientation support.