Assume ZoneView flush methods are called on the Swing thread now
Now that `ServerMessageHandler` and `ClientMessageHandler` are more careful about running on the Swing thread, they do
not cause off-thread flushes of the `ZoneView`. There are still a few places where off-thread acesss is possible (e.g.,
an app action that runs on a background thread) but these are far more rare and less likely to cause issues. Also
solving them all would be quite involved.