Update ServerMessageHandler to run mainly on the Swing thread
Most of `ServerMessageHandler`'s messages are able to interact with the model - or even the zone renderer directly -
which means they should be run on the Swing thread to avoid thread safety issues. As an added perk, now that these
handlers are serialied, we can avoid the need for the explicit mutex that was used to guard access to a portion of the
model.

`ClientMessageHandler` has also been updated to run more of its methods on the Swing thread. Most of the messages were
already handled this way, but there was a few that ought to run on the Swing thread but weren't.