Rewrite DebounceExecutor to minimize time spent between tasks
The previous implementation would always apply a full delay to any dispatch request, regardless of how long it had been
since the last scheduling of a task. This meant that execution A could complete, then execution B could be requested
_just before_ the end of the delay time and still be delayed the full time, resulting a double delay. This is noticeable
when events are coming in rapid succession (e.g., mouse dragging) as what should be a steady stream of task runs is
instead choppy and inconsistent.

The new implementation only delays tasks that have come in close succession to a previous task. If a long time has
passed, no delay is applied. The new implementation also avoids locks in favour of a lighter weight compare-exchange
operation.