Make almost all of MacroButtonProperties fields non-null
Most of the fields were either close to being non-nullable, or had getters that enforced non-nullability for callers. As
a result, most of them were readily made truly non-nullable, which enabled us to avoid the need for `null` checks
outside of setters and `readResolve()`.

`StringBuilder` is now used locally instead of `StringBuffer` in `modifySortString()` as it is lighterweight and there
is no multithreading concerns.