Modify the LightingComposite blender loops for better vectorization
The main change is to unroll the loops, assigning each color component's result to a separate variable. This makes it
clear to the JVM that the operations can be vectorized. On OpenJDK 17.0.1 there is minimal difference, but later
versions would not vectorize without this change.

`ConstrainedBrightenBlender` was also modified to use a simple conditional instead of a bit shift to decide between its
two cases. The result is a consistent 30% shaved off render times.