Flush aura cache when tokens with auras are changed
The `drawableAura` cache was optionally flushed based on whether the token currently has light sources or contributions
to an illumination. Auras don't interact with illumination, so this was missing tokens that used to have auras but now
have no light sources.

The new checks separate auras from normal lights. The aura cache is now cleared whenever the token is recognized as an
aura source, or just been given an aura.