Stop checking if it's okay to log messages
Calls to `Logger.isDebugEnabled()`, `Logger.isInfoEnabled()`, and `Logger.isTraceEnabled()` have been removed when they
simply gate a simple string log message, or a log message that could be parameterized.

There are a few instances left where we check prior to logging, specificaly if the thing we want to log is not readily
available and has some overhead in computing it. There are also some cases where `isDebugEnabled()` specifically is used
to enable secret features, and those uses still exist.