Implement a system for developer preferences
A new tab was added to the Preferences dialog that allows users to toggle developer features. These come with a warning
in the Preferences dialog since they are not intended to be used in typical situations. If the user leaves any options
enabled, they will also be shown a warning dialog the next time they start MapTool, so that they do not forget about the
enabled options or how to disable them.

The implementation is provided by a `DeveloperOptions.Toggle` enum that contains the four current developer
options. Each option wraps a `boolean` preference, allowing to get and set them, and provides translated labels and
descriptions for each. The enclosing class `DeveloperPreferences` provides a moethd to get all enabled options, which is
used to generate a warning to users. Every member of the enum is also automatically added to the Developer preferences
tab.

These are the four feature that used to be gated in `isDebugEnabled()`:
- `AutoSaveMeasuredInSeconds` to reinterpret the auto-save time as seconds instead of minutes
- `ShowPartitionDrawableBoundaries` to render spatial partition boundaries when rendering drawables
- `ShowAiDebugging` to show details on the A* algorithm
- `IgnoreGridShapeCache` to always recompute grid shapes whenever requested.

Meanwhile, `AbstractAStarWalker` has been reworked with regards to its debug labels, so that it clears only debug labels
instead of all labels. This is more important now that the option can be toggled without a restart.