Message dispatching is only needed for ClientConnection
`ServerConnection` was only ever used with one handler (`ServerMessageHandler`), which was set immediately after the
`ServerConnection` was created, and would not change through the lifetime of the `ServerConnection`. Additionally,
`AbstractServerConnection`'s nature as a `MessageHandler` was only for piping messages from `ClientConnection` into the
`ServerConnection`'s own dispatching mechanism.

So instead of relying on dispatching in `AbstractServerConnection`, we now require a `MessageHandler` to be passed at
the time of creation, and we attach it to each `ClientConnection` as they are created. This means only
`ClientConnection` need to know about message dispatching. So the registration methods could be moved out of
`Connection` into `ClientConnection`, and the dispatching implementation out of `AbstractConnection` into
`AbstractClientConnection`.