Move open()/close() methods out of Connection interface
Now `ServerConnection` and `ClientConnection` have their own `open()` and `close()` methods. In the case of
`ServerConnection`, the method was renamed to `start()` to indicate its purpose more clearly, as it is not related at
all to the opening of connections as in `ClientConnection`.

Both `ServerConnection` and `ClientConnection` now extend `AutoCloseable`, as their `close()` methods have the same
expectations.