Rename ClientConnection to Connection
Also the package name has been changed from `client` to `connection`.

This type represents a connection between a server and a client. It is to be used both server-side and client-side, and
so the use of the term "client" only distracts from its purpose. Similarly, the package name would suggest some
client-side utilities, but in reality contains implementations for connections.