Separate darkness rendering from light rendering for non-GM players
Non-GM players need to have darkness rendered as blackness rather than as lights. We used to accomplish this by
rendering each darkness as a regular light (as the GM would see it), then would paint each darkness as black
afterwards.

Now we don't even represent the darkness as a `DrawableLight` unless the player is a GM. This way we aren't wasting time
compositing darkness lights that won't even be visible in the end. We also don't render the darknesses individually, but
instead render the entire darkened area as black in a single pass.