Ensure PackedFile is always closed and remove PackedFile.finalize()
Almost all `PackedFile` instances are wrapped in a try-with-resources to ensure they are always closed. The one
exception guarantees the instance is closed in a `finally` block.

The following places used to (at least partially) rely on PackedFile.finalize() but now use try-with-resources:
- `ImageFileImagePanelModel.getDecorations()`
- `PersistenceUtils.loadCampaignProperties()`
- `PackedFile`

The following place used to call `PackedFile.close()` in a `finally` block now uses try-with-resources:
- `PersistenceUtil.loadCampaign()`

The following place still uses a `finally` block because there is extra logic surrounding the close() call:
- `PersistenceUtil.saveCampaign()`