Remove workaround for #1976
The workaround was needed as old overlay contents were not being erased when the page background was transparent. That
problem no longer seems to exist, so the workaround has been removed in favour of always setting a fully transparent
background. This also seems to fix the blue flicker that has plagued overlays.