Drive the layer selection dialog from tools
Instead of the dialog having to discover the current renderer, we now have the tools directly set the layer it should
present as selected. This removes an unnecessary assumption in `LayerSelectionDialog` while also making it clear in
`DefaultTool` that the dialog is being updated to use the same layer expected by the tool.