Fix double title bars with grouped floating frames
A new `DialogFloatingContainer` implementation (`SingleTitleBarDialogFloatingContainer`) is provided which always leaves
itself undecorated, but which will decorate the root pane with dialog decorations when needed. The border is also always
set to a resize border rather than sometimes having it and sometimes not. A new `DockingManager`
implementation (`MapToolDockingManager`) is needed just to instantiate `SingleTitleBarDialogFloatingContainer`.