Add optional token list parameter to getIllumination()
New usage: `getIllumination(x, y, mapNameOrId, tokenIds)`

The token list controls allows the player view to be customized. It must be a JSON array of tokens IDs for the current
zone, with any unknown tokens resulting in an error.

If not provided, the player view will be the current player view for the map.

If the list of token IDs is empty, the `PlayerView` will use a token view but with no tokens involved, i.e., no token
sight will be accounted for.