Add macro function getIllummination()
Usage: `getIllumination(x, y)`

This basic form of the function uses the current `PlayerView` for the map, so the lumens value returned will match what
the lumens overlay shows on the client at the given point. Personal lights and sight magnifiers will be accounted for if
they are in the current view.