Convert MTScript input numbers and output objects
MTScript numbers are BigDecimals, so require conversion to a number type
that graal Values can contain.

Objects returned from JavaScript functions are PolyglotMaps,
which are Host Objects instead of subclasses of Value
and get turned into a HashMap,
which needs to be serialized to JSON to be usable from MTScript.

This allows