Implement light sources on tokens
It is now possible for tokens to carry light source definitions, much like campaigns can. These new light sources are
called "unique lights" as they are only available to the token that defines them. If any unique light source are defined
on a token, they can be accessed like any other lights via the right-click menu, under the special category "Unique".

In order to support lookup of such light sources, `AttachedLightSource#resolve()` now also accepts a `Token` for
context, in addition to the `Campaign`. The token is checked first for a matching light source, then the campaign is
checked as a fallback.

For networking, the `TokenPropertyValueDto.lightSource` field was added back in because creating a unique light source
now requires the complete definition to be passed.

This commit does not add any way for a user to add any light sources to token. That will come later.