Fix to not hide partially GM-only light sources in the menu
It's possible for only some lights in an aura light source to be marked as GM-only (or OWNER-only). These were not being
shown in the right-click menu despite players being able to see them once equipped. Now these auras will show in the
menu, and will only be hidden if the entire auras is GM-only.

This also include a refactor to avoid duplicated logic between campaign light sources and unique light sources, while
also avoiding some of the complicated looping.