Remove use of tika for stack traces
`ServerMessageHandler` is unique in using tika to generate stack traces for logging and for displaying to the user. In
addition, `MapTool.showError()` already logs the error (both message and exception). So now we no longer bother with
tika here, nor do we log the exception here, but instead leave it as a responsibility for `MapTool.showError()`.

`MapTool.showError()` also now uses `I18N.getText()` instead of `I18N.getString()` so that the message key is shown
instead of `null` in the case of a missing translation.