Make Zone.Layer enum immutable
It is now the responsibility of `ZoneRenderer` to decide whether to render a layer. This keeps the decision local to the
current map being rendered rather than being globally applied. In the future we may even be able to use a dedicated
rednerer in the `ExportDialog` to avoid any sullying of export screenshots with the regular application rendering.