Reimplement TokenPanelTreeModel.View as a static class
Now that `View` is no longer an enum, it is possibly to create instances dynamically according to the `Zone.Layer`
constants that exist. Any new non-token layers will be included in the GUI, but more particular additions will require
some updates to this logic.

Alonside that change is a fulfillment of a long-standing TOOD note. The `View` and `TokenFilter` are more coupled now,
with `TokenFilter` providing all filtering functionality and configuration, while `View` provides a display name and
acts as a lookup key. The `GMFilter`, `ObjectFilter`, and `BackgroundFilter` have all been merged into a single
`AdminLayerFilter` that enforces GM and layer checks. The other filters (`PlayerTokenFilter`, `NPCTokenFilter`, and
`LightSourceFilter`) remain essentially as they are, just updated for the new system and for consistency.

Related to that is some minor refactoring:
- The `View.isRequired` flag is gone now as it was always `false`.
- The population of `viewMap` was moved out of `TokenFilter` and into `updateInternal()`. This allows coalescing most of
  the `updateInternal()` loops into a single small loop responsible for populating `viewMap` and `currentViewList`. This
  also allows us to make the view lists unmodifiable to ensure no one else is messing with the model's internal state.
- The `TreeModelEvent` created for structure changes does not need the children and indices parameters to be
  provided. Removing them avoids any need to keep around static `View` instances.