Reimplement ExportDialog.ExportLayers as a static class
By moving away from an enum, it is now possible to create `ExportLayers` instances dynamically based on whatever
`Zone.Layer` constants exist, and we now do exactly that.

This does not address the GUI aspect of `ExportDialog`. If new layers are defined, it will still be necessary to
manually add checkboxes to the dialog view in order to control the layers.