Add Layer.isTokenLayer()
This new method acts as a property of layers, stating whether or not the layer holds tokens. In order to return true,
the layer must contain tokens as opposed to stamps. The property should be used in situations where the nature as tokens
is important, rather than - for example - the ability to see or pathfind.