Replace configSentryRelease task with processResources configuration
This avoids dirtying the `src/` directory with build artifacts, thus introducing unexpected dependencies between
tasks. The `:run` task naturally no longer depends on `:spotlessCheck` in any way.

The `src/main/resources/sentry.properties` file is no longer ignored by git as it should not be generated by the build
anymore. If the file is present in a dev tree, it should be deleted.