Make light source detection more consistent
Since players and GMs disagree on which light sources exist, and since GMs can enable "player view", we need to be able
to track light sources based on the current role. This has implications for illumination as well, so we now key the
light source map and illuminations cache by player role in addition to the existing keys. Doing so allows
`lightSourceMap` to always be `ZoneView`'s source of truth when it comes to light sources, and it works even in the
presence of view changes.

On a related note, auras are now only filtered in `getDrawableAuras()`, but that method is now dependent on the current
view. This makes it so that, e.g., GM-only auras are not returned  when the GM is in player view.

There is now one place (`updateLightSourcesFromTokens()`) that checks whether or not a token should be added to the
light source map. This avoids inconsistencies in how tokens are handled at campaign load vs being added or modified
later. The event handlers that need to use it have also been updated to avoid any repeated logic.

With the other changes, `MapTool.getPlayer().isEffectiveGM()` is no longer needed since we always work off the provided
view rather than trying to determine it ourselves. The same logic, if needed again, can be attained by checking
`getEffectiveRole()` instead.