Fix look up of event macros to respect ownership
We had checks to ensure that lib:tokens owned by any players did not have their onCampaignLoad even executed. But the
check didn't `continue` the correct loop, so outside of `Token.isOwnedByAll()`, it had no effect. This changes that to
use `Token.isOwnedByAny()`, and now the right loop is naturally manipulated.

Also removes some unecessary nesting of blocks and an unused method.