Rework walker slightly to not depend on global information.
Instead of getting the current zone renderer, we now get the renderer for the zone that the walker was created
with. This should in practice be the same thing, but reduces the walker's dependence on global knowledge and avoids
potential edge cases.

Also uses that renderer's player view rather than directly checking if the player is effectively a GM. The check can
also be done when building the FOW geometry rather than on every potential cell step.