Add Facing enum for easier reading of topology edge facing
Instead of an ambiguous boolean to indicate "front side", or elsewhere "facing away", we now indicate facing using the
`Facing` enum. This enum is based on the fact that each line segment will have an island on one side and an ocean on the
other side, thus allowing names that indicate the position of the origin point as being either on the island side or the
ocean side.