Simplify AreaTree and related types
The quartet of `AreaTree`, `AreaContainer`, `AreaIsland`, and `AreaOcean` have been condensed into `AreaTree` for a
clearer structure and smaller code size. `AreaTree` no longer allows modifications after construction, has a more
efficient construction process using JTS' `Polygonizer`, and more directly supports the patterns established in
cbf91d1c0. Each node in the tree is now represented by a simple `AreaTree.Node` type which only contains references to
its children and the associated `AreaMeta`. All logic is either in `AreaTree` (for traversing the tree) or in
`AreaMeta` (for operations on the region itself).

`AreaMeta` has been simplified as well. It is no longer mutable, and doesn't possess or build an AWT `Area`. It is
represented as a list of vertices which defines the polygonal boundary of the region it represents. There are other
fields, but they are derived entirely from the vertices and exist solely for efficiency. `AreaMeta` is also now capable
of representing the root ocean, which simplifies the usage of `AreaTree` nodes as they no longer have to check for
possibly-null `AreaMeta` for just this one case.

`AreaTree.getArea()` has been removed since it only existed to support `VisibilityInspector`. Since that is just a
debugging tool, and the method is kind of dangerous (doesn't protect against changes), it doesn't make sense to keep it
around. Instead `VisibilityInspector` keeps track of its own `Area` objects, alongside some other minor
simplifications.