Avoid unnecssary intersection calculations in visibility sweep
This change moves much closer to Asano's original algorithm as described in section 3.2 of "Efficient Computation of
Visibility Polygons" ([arXiv:1403.3905](https://arxiv.org/abs/1403.3905)). The critical difference is in how we find the
nearest wall, in particular:
1. Keeping the open walls ordered by distance to the origin, so that the nearest one is always at the front.
2. Using a cheaper orientation-based comparator for the open walls that does not depend on any finding any
   intersections.

Since we don't have to intersect walls in order to find the nearest one, we can now limit ourselves to a single nearest
wall check per iteration, as opposed to the two that used to be used. Intersections are only calculated when we need to
add a point to the result that does not lie on one of the input endpoints.