Establish patterns for all VBL types in VisionBlockingAccumulator
Given that topology trees are alternating islands and oceans, and since vision is only blocked at boundaries of each,
all VBL types follow a common pattern in their case work:
1. Find containers that vision is allowed to pass through (oceans for Wall and Cover, islands for Hill and Pit).
2. Use part of those containers' boundaries to block vision.
3. Use part of those containers' childrens' boundaries to block vision.

The new `blockVisionBeyondContainer(AreaIsland)` and `blockVisionBeyondContainer(AreaOcean)` helpers give direct support
to this pattern and reduces noise in the case work. The `AreaIsland` overload was previously called
`addIslandForHillBlocking(AreaIsland)` and used for Hill and Pit VBL, but the new naming makes it clearer that it is not
specific to Hill VBL and is consistent with the `AreaOcean` version.

Another pattern relies on the fact that the result of `AreaTree.getDeepestContainerAt(Point2D)` is either an ocean or
has an ocean as a parent (the same is not true of islands). The case work for Hill and Cover VBL is identical in both
cases, so they can be flattened by anchoring the implemetations at this nearest ocean then mixing in the presence of
islands as needed.

As a convenience, it is now possible to call `VisionBlockingAccumulator.add(TopologyType, AreaTree)` to add topology of
a given type, rather than having separate public methods for each. This makes `FogUtil` a bit more natural to read.